requirements.txt
discord.py>=2.3.0
openai>=1.0.0
pyyaml>=6.0
httpx>=0.24.0